make CV_64FC1 or CV_64FC3 default

+++ inio/gamma_correction.hpp
@@ -2,4 +2,4 @@
 
 #include <opencv2/opencv.hpp>
 
-cv::Mat gc(cv::Mat src, float ganma);
+cv::Mat gc(cv::Mat src, double ganma);
